+++ examples/test_mount.rs
@@ -27,8 +27,8 @@ const FILE_TO_PRINT: &'static str = "README.TXT";
 const FILE_TO_CHECKSUM: &'static str = "64MB.DAT";
 
 use embedded_sdmmc::{
-    Block, BlockCount, BlockDevice, BlockIdx, Controller, Error, Mode, TimeSource, Timestamp,
-    VolumeIdx,
+    Block, BlockCount, BlockDevice, BlockIdx, Error, Mode, TimeSource, Timestamp, VolumeIdx,
+    VolumeManager,
 };
 use std::cell::RefCell;
 use std::fs::File;
@@ -121,14 +121,14 @@ fn main() {
         .map_err(Error::DeviceError)
         .unwrap();
     println!("lbd: {:?}", lbd);
-    let mut controller = Controller::new(lbd, Clock);
+    let mut volume_mgr = VolumeManager::new(lbd, Clock);
     for i in 0..=3 {
-        let volume = controller.get_volume(VolumeIdx(i));
+        let volume = volume_mgr.get_volume(VolumeIdx(i));
         println!("volume {}: {:#?}", i, volume);
         if let Ok(mut volume) = volume {
-            let root_dir = controller.open_root_dir(&volume).unwrap();
+            let root_dir = volume_mgr.open_root_dir(&volume).unwrap();
             println!("\tListing root directory:");
-            controller
+            volume_mgr
                 .iterate_dir(&volume, &root_dir, |x| {
                     println!("\t\tFound: {:?}", x);
                 })
@@ -137,15 +137,15 @@ fn main() {
             println!(
                 "\tFound {}?: {:?}",
                 FILE_TO_PRINT,
-                controller.find_directory_entry(&volume, &root_dir, FILE_TO_PRINT)
+                volume_mgr.find_directory_entry(&volume, &root_dir, FILE_TO_PRINT)
             );
-            let mut f = controller
+            let mut f = volume_mgr
                 .open_file_in_dir(&mut volume, &root_dir, FILE_TO_PRINT, Mode::ReadOnly)
                 .unwrap();
             println!("FILE STARTS:");
             while !f.eof() {
                 let mut buffer = [0u8; 32];
-                let num_read = controller.read(&volume, &mut f, &mut buffer).unwrap();
+                let num_read = volume_mgr.read(&volume, &mut f, &mut buffer).unwrap();
                 for b in &buffer[0..num_read] {
                     if *b == 10 {
                         print!("\\n");
@@ -155,43 +155,43 @@ fn main() {
             }
             println!("EOF");
             // Can't open file twice
-            assert!(controller
+            assert!(volume_mgr
                 .open_file_in_dir(&mut volume, &root_dir, FILE_TO_PRINT, Mode::ReadOnly)
                 .is_err());
-            controller.close_file(&volume, f).unwrap();
+            volume_mgr.close_file(&volume, f).unwrap();
 
-            let test_dir = controller.open_dir(&volume, &root_dir, "TEST").unwrap();
+            let test_dir = volume_mgr.open_dir(&volume, &root_dir, "TEST").unwrap();
             // Check we can't open it twice
-            assert!(controller.open_dir(&volume, &root_dir, "TEST").is_err());
+            assert!(volume_mgr.open_dir(&volume, &root_dir, "TEST").is_err());
             // Print the contents
             println!("\tListing TEST directory:");
-            controller
+            volume_mgr
                 .iterate_dir(&volume, &test_dir, |x| {
                     println!("\t\tFound: {:?}", x);
                 })
                 .unwrap();
-            controller.close_dir(&volume, test_dir);
+            volume_mgr.close_dir(&volume, test_dir);
 
             // Checksum example file. We just sum the bytes, as a quick and dirty checksum.
             // We also read in a weird block size, just to exercise the offset calculation code.
-            let mut f = controller
+            let mut f = volume_mgr
                 .open_file_in_dir(&mut volume, &root_dir, FILE_TO_CHECKSUM, Mode::ReadOnly)
                 .unwrap();
             println!("Checksuming {} bytes of {}", f.length(), FILE_TO_CHECKSUM);
             let mut csum = 0u32;
             while !f.eof() {
                 let mut buffer = [0u8; 2047];
-                let num_read = controller.read(&volume, &mut f, &mut buffer).unwrap();
+                let num_read = volume_mgr.read(&volume, &mut f, &mut buffer).unwrap();
                 for b in &buffer[0..num_read] {
                     csum += u32::from(*b);
                 }
             }
             println!("Checksum over {} bytes: {}", f.length(), csum);
-            controller.close_file(&volume, f).unwrap();
+            volume_mgr.close_file(&volume, f).unwrap();
 
-            assert!(controller.open_root_dir(&volume).is_err());
-            controller.close_dir(&volume, root_dir);
-            assert!(controller.open_root_dir(&volume).is_ok());
+            assert!(volume_mgr.open_root_dir(&volume).is_err());
+            volume_mgr.close_dir(&volume, root_dir);
+            assert!(volume_mgr.open_root_dir(&volume).is_ok());
         }
     }
 }