+++ examples/write_test.rs
@@ -18,8 +18,8 @@ extern crate embedded_sdmmc;
 const FILE_TO_WRITE: &str = "README.TXT";
 
 use embedded_sdmmc::{
-    Block, BlockCount, BlockDevice, BlockIdx, Controller, Error, Mode, TimeSource, Timestamp,
-    VolumeIdx,
+    Block, BlockCount, BlockDevice, BlockIdx, Error, Mode, TimeSource, Timestamp, VolumeIdx,
+    VolumeManager,
 };
 use std::cell::RefCell;
 use std::fs::{File, OpenOptions};
@@ -118,14 +118,14 @@ fn main() {
         .map_err(Error::DeviceError)
         .unwrap();
     println!("lbd: {:?}", lbd);
-    let mut controller = Controller::new(lbd, Clock);
+    let mut volume_mgr = VolumeManager::new(lbd, Clock);
     for volume_idx in 0..=3 {
-        let volume = controller.get_volume(VolumeIdx(volume_idx));
+        let volume = volume_mgr.get_volume(VolumeIdx(volume_idx));
         println!("volume {}: {:#?}", volume_idx, volume);
         if let Ok(mut volume) = volume {
-            let root_dir = controller.open_root_dir(&volume).unwrap();
+            let root_dir = volume_mgr.open_root_dir(&volume).unwrap();
             println!("\tListing root directory:");
-            controller
+            volume_mgr
                 .iterate_dir(&volume, &root_dir, |x| {
                     println!("\t\tFound: {:?}", x);
                 })
@@ -134,14 +134,14 @@ fn main() {
             // This will panic if the file doesn't exist, use ReadWriteCreateOrTruncate or
             // ReadWriteCreateOrAppend instead. ReadWriteCreate also creates a file, but it returns an
             // error if the file already exists
-            let mut f = controller
+            let mut f = volume_mgr
                 .open_file_in_dir(&mut volume, &root_dir, FILE_TO_WRITE, Mode::ReadOnly)
                 .unwrap();
             println!("\nReading from file {}\n", FILE_TO_WRITE);
             println!("FILE STARTS:");
             while !f.eof() {
                 let mut buffer = [0u8; 32];
-                let num_read = controller.read(&volume, &mut f, &mut buffer).unwrap();
+                let num_read = volume_mgr.read(&volume, &mut f, &mut buffer).unwrap();
                 for b in &buffer[0..num_read] {
                     if *b == 10 {
                         print!("\\n");
@@ -150,17 +150,17 @@ fn main() {
                 }
             }
             println!("EOF\n");
-            controller.close_file(&volume, f).unwrap();
+            volume_mgr.close_file(&volume, f).unwrap();
 
-            let mut f = controller
+            let mut f = volume_mgr
                 .open_file_in_dir(&mut volume, &root_dir, FILE_TO_WRITE, Mode::ReadWriteAppend)
                 .unwrap();
 
             let buffer1 = b"\nFile Appended\n";
             let buffer = [b'a'; 8192];
             println!("\nAppending to file");
-            let num_written1 = controller.write(&mut volume, &mut f, &buffer1[..]).unwrap();
-            let num_written = controller.write(&mut volume, &mut f, &buffer[..]).unwrap();
+            let num_written1 = volume_mgr.write(&mut volume, &mut f, &buffer1[..]).unwrap();
+            let num_written = volume_mgr.write(&mut volume, &mut f, &buffer[..]).unwrap();
             println!("Number of bytes written: {}\n", num_written + num_written1);
 
             f.seek_from_start(0).unwrap();
@@ -168,12 +168,12 @@ fn main() {
             println!(
                 "\tFound {}?: {:?}",
                 FILE_TO_WRITE,
-                controller.find_directory_entry(&volume, &root_dir, FILE_TO_WRITE)
+                volume_mgr.find_directory_entry(&volume, &root_dir, FILE_TO_WRITE)
             );
             println!("\nFILE STARTS:");
             while !f.eof() {
                 let mut buffer = [0u8; 32];
-                let num_read = controller.read(&volume, &mut f, &mut buffer).unwrap();
+                let num_read = volume_mgr.read(&volume, &mut f, &mut buffer).unwrap();
                 for b in &buffer[0..num_read] {
                     if *b == 10 {
                         print!("\\n");
@@ -182,10 +182,10 @@ fn main() {
                 }
             }
             println!("EOF");
-            controller.close_file(&volume, f).unwrap();
+            volume_mgr.close_file(&volume, f).unwrap();
 
             println!("\nTruncating file");
-            let mut f = controller
+            let mut f = volume_mgr
                 .open_file_in_dir(
                     &mut volume,
                     &root_dir,
@@ -195,20 +195,20 @@ fn main() {
                 .unwrap();
 
             let buffer = b"Hello\n";
-            let num_written = controller.write(&mut volume, &mut f, &buffer[..]).unwrap();
+            let num_written = volume_mgr.write(&mut volume, &mut f, &buffer[..]).unwrap();
             println!("\nNumber of bytes written: {}\n", num_written);
 
             println!("\tFinding {}...", FILE_TO_WRITE);
             println!(
                 "\tFound {}?: {:?}",
                 FILE_TO_WRITE,
-                controller.find_directory_entry(&volume, &root_dir, FILE_TO_WRITE)
+                volume_mgr.find_directory_entry(&volume, &root_dir, FILE_TO_WRITE)
             );
             f.seek_from_start(0).unwrap();
             println!("\nFILE STARTS:");
             while !f.eof() {
                 let mut buffer = [0u8; 32];
-                let num_read = controller.read(&volume, &mut f, &mut buffer).unwrap();
+                let num_read = volume_mgr.read(&volume, &mut f, &mut buffer).unwrap();
                 for b in &buffer[0..num_read] {
                     if *b == 10 {
                         print!("\\n");
@@ -217,7 +217,7 @@ fn main() {
                 }
             }
             println!("EOF");
-            controller.close_file(&volume, f).unwrap();
+            volume_mgr.close_file(&volume, f).unwrap();
         }
     }
 }